updated messages & spectogram output when using APEX

+++ signaltrain/io_methods.py
@@ -481,16 +481,16 @@ def plot_spectrograms(model, mag_val, mag_val_hat):
 	savefig('mag_hat.png')
 
 	#if isinstance(model, nn_proc.AsymMPAEC):     # Plot the spectrograms
-	plt.matshow(model.mpaec.dft_analysis.conv_analysis_real.weight.data.cpu().numpy()[:, 0, :] + 1)
+	plt.matshow(model.mpaec.dft_analysis.conv_analysis_real.weight.data.cpu().numpy().astype(float)[:, 0, :] + 1)
 	plt.title('Conv-Analysis Real')
 	savefig('conv_anal_real.png')
-	plt.matshow(model.mpaec.dft_analysis.conv_analysis_imag.weight.data.cpu().numpy()[:, 0, :])
+	plt.matshow(model.mpaec.dft_analysis.conv_analysis_imag.weight.data.cpu().numpy().astype(float)[:, 0, :])
 	plt.title('Conv-Analysis Imag')
 	savefig('conv_anal_imag.png')
-	plt.matshow(model.mpaec.dft_synthesis.conv_synthesis_real.weight.data.cpu().numpy()[:, 0, :])
+	plt.matshow(model.mpaec.dft_synthesis.conv_synthesis_real.weight.data.cpu().numpy().astype(float)[:, 0, :])
 	plt.title('Conv-Synthesis Real')
 	savefig('conv_synth_real.png')
-	plt.matshow(model.mpaec.dft_synthesis.conv_synthesis_imag.weight.data.cpu().numpy()[:, 0, :])
+	plt.matshow(model.mpaec.dft_synthesis.conv_synthesis_imag.weight.data.cpu().numpy().astype(float)[:, 0, :])
 	plt.title('Conv-Synthesis Imag')
 	savefig('conv_synth_imag.png')
